+++ exercises/2/webapp/src/components/NewsItem.vue
